Add support to use GLFW (linux only for now);
This implements a very thin and hacky layer that makes glfw behave as
SDL, so to avoid a major rewrite of RecastDemo main.

I've added a premake option "--use-glfw" with a configuration for linux
only, adding support to windows and macosx should be trivial and need
only changes to the premake file. I'm not doing this right now due to
not being able to test it on these platforms.

Also note that the premake file changes may not be optimal, I still do not
have a good premake knowledge.