Fix for wrongly simplified neighbour contours
Two fixes for walkContour/simplifyContour:
- Fix for floating error innacuracy when calculating the maximum
deviation during simplifyContour: ensure points are passed to
distancePtSeg on the same order for both neighbour contours.
- Fix for regions touching only in one vertex: during walkContour mark
vertices that touch another region in only one vertex, and then let
simplifyContour insert that vertex in the initial polygon.

See the post "Two bugs in contour generation (and fixes)" in Recast
forum for a detailed description
https://groups.google.com/forum/#!topic/recastnavigation/OHK5JWLvWs0