Add parent indices in Dijkstra searches
This changes the findPolysAroundCircle, findPolysAroundShape and
findLocalNeighbourhood functions to return indices into the poly array
for the parents, instead of returning the polygon references. This makes
it much more convenient to retrace paths after performing the searches,
based on the parent arrays.

Also make some arguments to the aforementioned functions required as the
APIs are not very useful without them. Update documentation to specify
this and perform addition parameter validation.

Fix #206