Move FastLZ to it's own project in premake5 config
so we don't get errors about certain flags being C++ only.  Specifically we get this error when compiling on Ubuntu with gcc:

cc1: error: command-line option ‘-Wno-class-memaccess’ is valid for C++/ObjC++ but not for C [-Werror]

This isn't a terribly aesthetic solution to this (yet another project in the workspace) but it works and is clean.