Docs Overhaul (#680)
A lot of the auxiliary project documentation was pretty old, redundant, or could use some polishing.  This change fixes up some wording, grammar, layout and generally cleans up the docs a bit.  It's not perfection, but it's an improvement

I also tried to streamline and beautify the top-level `README.md` to be more appealing, clear, and to direct people to the official docs.  To this end, I moved the building instructions to This is based on what I've seen in some other popular github projects.