Explicitly use html tags in Changelog to work around bug in doxygen
There's a bug in doxygen's markdown html generator when specifying header text that's just link text.  This is fixed in newer versions of doxygen, but the ubuntu images in github actions runners don't have a new enough version to work around this.  Instead, just write out the html explicitly for now.