Add areaID to add obstacle calls
Obstacles typically write DT_TILECACHE_NULL_AREA into the navmesh structures, which results in a hole being added.  Now we can pass in any area ID which will instead mark the triangles.  Current use case it to mark the area under a pod-spawner as DO-NOT-ENTER.  This allows the zombies to be spawned 'inside' the pod collision and can navigate out of the zone.