Explicitly set c++ version per-project.  Set RecastDemo and Tests projects to C++20
We don't care as much about ultra-compatibility with RecastDemo or Tests.  C++20 seems fairly universal at this point considering Unreal 5 ships with it as the default requirement.

We want to keep the other stuff to C++98 still because eventaully we'll convert it to C99 and not having a ton of modern C++ features everywhere make it significantly easier.